checking if inputs are coordinates, resetting error messages
