+++ src/components/SearchBar.vue
@@ -39,6 +39,7 @@
 <script setup lang="ts">
 import { computed, ref, watch } from "vue";
 import { useSearchStore } from "src/stores/search-store";
+import isCoordinate from "./is-coordinate";
 
 const searchStore = useSearchStore();
 const searchTerm = ref("");
@@ -50,20 +51,6 @@ const searchBarBackground = computed(() => {
 
 function clearSearchTerm() {
   searchTerm.value = "";
-}
-
-function isCoordinate(input: string) {
-  const coordRegex = /^\s*(-?\d+(\.\d+)?)[,\s]+(-?\d+(\.\d+)?)\s*$/;
-  const match = input.match(coordRegex);
-  if (!match || !match[1] || !match[3]) return null;
-
-  const lat = parseFloat(match[1]);
-  const lng = parseFloat(match[3]);
-
-  if (lat >= -90 && lat <= 90 && lng >= -180 && lng <= 180) {
-    return { lat, lng };
-  }
-  return null;
 }
 
 watch(searchTerm, async (value) => {
