+++ src/css/app.scss
@@ -54,7 +54,7 @@ h2 {
   width: 100%;
   display: flex;
   flex-direction: column;
-  gap: 24px;
+  gap: 8px;
 }
 
 .input-container {
